fix how to check for tests if standard-test-roles is not installed
when STR is not installed it is difficult to check for tag as in yaml it can be defined in different ways

in that case just do basic test if there is valid test playbook.

It is recommended to run this function again on Jenkins slave with STR installed.

This should fix https://pagure.io/fedora-ci/general/issue/91